+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Description of changes by version
 
+# v0.1.3
+- Added support for multiple prefixes/suffixes.
+- Added the "prefixchange" and "suffixchange" events.
+- Added complete documentation for each function.
+- Added function affixValue.
+- Fixed some minor user experience issues.
+
 # v0.1.2
 - Fixed pasting event handlers.
 - Added handlers for context menu interaction.